Implemented and made changes in user authorization code

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,7 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
+using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
@@ -10,12 +14,20 @@ namespace RestaurantAPI.Services
 		private readonly RestaurantDbContext _dbContext;
 		private readonly IMapper _mapper;
 		private readonly ILogger<RestaurantService> _logger;
+		private readonly IAuthorizationService _authorizationService;
+		private readonly IUserContextService _userContextService;
 
-		public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+		public RestaurantService(RestaurantDbContext dbContext,
+								 IMapper mapper,
+								 ILogger<RestaurantService> logger,
+								 IAuthorizationService authorizationService,
+								 IUserContextService userContextService)
 		{
 			_dbContext = dbContext;
 			_mapper = mapper;
 			_logger = logger;
+			_authorizationService = authorizationService;
+			_userContextService = userContextService;
 		}
 
 		public IEnumerable<RestaurantDto> GetAll()
@@ -48,6 +60,8 @@ namespace RestaurantAPI.Services
 		{
 			var restaurant = _mapper.Map<Restaurant>(createRestaurantDto);
 
+			restaurant.CreatedById = _userContextService.UserId;
+
 			_dbContext.Restaurants.Add(restaurant);
 			_dbContext.SaveChanges();
 
@@ -61,6 +75,13 @@ namespace RestaurantAPI.Services
 				.FirstOrDefault(r => r.Id == id);
 
 			if (restaurant is null) throw new NotFoundException("Restaurant not found");
+
+			var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+			if (!authResult.Succeeded)
+			{
+				throw new ForbiddenException();
+			}
 
 			restaurant.Name = updateRestaurantDto.Name;
 			restaurant.Description = updateRestaurantDto.Description;
@@ -76,6 +97,13 @@ namespace RestaurantAPI.Services
 				.FirstOrDefault(r => r.Id == id);
 
 			if (restaurant is null) throw new NotFoundException("Restaurant not found");
+
+			var authResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+			if (!authResult.Succeeded)
+			{
+				throw new ForbiddenException();
+			}
 
 			_dbContext.Restaurants.Remove(restaurant);
 			_dbContext.SaveChanges();
